requirements.txt
Jinja2>=3.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.1.8

langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.20
langchain-text-splitters>=0.0.1

# torch>=2.0.0
sentence-transformers==2.2.2
transformers==4.21.0        
llama-cpp-python>=0.2.0
accelerate>=0.25.0
numpy==1.26.4
huggingface_hub==0.16.4

# Utilities
python-dotenv>=1.0.0  
pandas>=1.5.0         
aiofiles>=23.1.0      
psutil>=5.9.0                   

#File Loaders for database.py ingest (uncomment if needed)
pypdf>=3.15.0
python-docx>=1.1.0
unstructured>=0.10.0
# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.7 

# Data validation
pydantic==2.5.3
pydantic[email]
itsdangerous>=2.0.0 

# --- CORRECTED LINES ---
Authlib>=1.0.0 
httpx>=0.24.0 
starlette>=0.35.0 